+++ src/main/java/ru/hogwarts/school/controller/StudentController.java
@@ -33,7 +33,7 @@ public class StudentController {
     }
 
     @GetMapping("{id}")
-    public StudentDTO getStudent(@PathVariable Integer id) {
+    public StudentDTO getStudent(@PathVariable long id) {
         return studentService.getStudentById(id);
     }
 
@@ -43,12 +43,12 @@ public class StudentController {
     }
 
     @PutMapping("/{id}")
-    public StudentDTO updateStudent(@PathVariable("id") int id, @RequestBody StudentDTO dto) {
+    public StudentDTO updateStudent(@PathVariable("id") long id, @RequestBody StudentDTO dto) {
         return studentService.updateStudent(dto);
     }
 
     @DeleteMapping("/{id}")
-    public void removeStudent(@PathVariable("id") int id) {
+    public void removeStudent(@PathVariable("id") long id) {
         studentService.removeStudent(id);
     }
 
@@ -68,7 +68,7 @@ public class StudentController {
     }
 
     @GetMapping("/getStudentsByFaculty")
-    public List<StudentDTO> getStudentsByIdFaculty(int id) {
+    public List<StudentDTO> getStudentsByIdFaculty(long id) {
         return studentService.getStudentsByIdFaculty(id);
     }
 
@@ -89,7 +89,7 @@ public class StudentController {
     }
 
     @GetMapping("/{id}/avatar/preview")
-    public ResponseEntity<byte[]> downloadAvatar(@PathVariable int id) {
+    public ResponseEntity<byte[]> downloadAvatar(@PathVariable long id) {
         Avatar avatar = avatarService.findAvatar(id);
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.parseMediaType(avatar.getMediaType()));
@@ -98,7 +98,7 @@ public class StudentController {
     }
 
     @GetMapping("/{id}/avatar")
-    public void downloadAvatar(@PathVariable int id, HttpServletResponse response) throws IOException {
+    public void downloadAvatar(@PathVariable long id, HttpServletResponse response) throws IOException {
         Avatar avatar = avatarService.findAvatar(id);
 
         Path path = Path.of(avatar.getFilePath());
